move mocks, test for create_user
