+++ pkg/cmd/add_repo_test.go
@@ -2,15 +2,7 @@ package cmd
 
 import (
 	"testing"
-
-	"github.com/ZupIT/ritchie-cli/pkg/formula"
 )
-
-type repoAdder struct{}
-
-func (repoAdder) Add(d formula.Repository) error {
-	return nil
-}
 
 func TestNewAddRepoCmd(t *testing.T) {
 	cmd := NewAddRepoCmd(repoAdder{}, inputTextMock{}, inputURLMock{}, inputIntMock{})
