+++ pkg/cmd/autocomplete_test.go
@@ -2,8 +2,6 @@ package cmd
 
 import (
 	"testing"
-
-	"github.com/ZupIT/ritchie-cli/pkg/autocomplete"
 )
 
 func TestNewAutocompleteCmd(t *testing.T) {
@@ -16,12 +14,6 @@ func TestNewAutocompleteCmd(t *testing.T) {
 	if err := cmd.Execute(); err != nil {
 		t.Errorf("%s = %v, want %v", cmd.Use, err, nil)
 	}
-}
-
-type autocompleteGenMock struct{}
-
-func (autocompleteGenMock) Generate(s autocomplete.ShellName) (string, error) {
-	return "autocomplete", nil
 }
 
 func TestNewAutocompleteZsh(t *testing.T) {
