+++ pkg/cmd/create_formula_test.go
@@ -4,12 +4,6 @@ import (
 	"testing"
 )
 
-type formCreator struct{}
-
-func (formCreator) Create(formulaCmd string) error {
-	return nil
-}
-
 func TestNewCreateFormulaCmd(t *testing.T) {
 	cmd := NewCreateFormulaCmd(formCreator{}, inputTextMock{})
 	if cmd == nil {
