+++ pkg/cmd/create_user_test.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+	"testing"
+)
+
+func TestNewCreateUserCmd(t *testing.T) {
+	cmd := NewCreateUserCmd(userManagerMock{}, inputTextMock{}, inputEmailMock{}, inputPasswordMock{})
+	if cmd == nil {
+		t.Errorf("NewCreateUserCmd got %v", cmd)
+
+	}
+
+	if err := cmd.Execute(); err != nil {
+		t.Errorf("%s = %v, want %v", cmd.Use, err, nil)
+	}
+}
